release from fill-buffer will now account for dirty line
if the line being replaced is dirty then, in the first cycle a read request for the same line is presented, in the next cycle the dirty line is sent to the memory and the new line is writen. In the last cycle the latest read from the core is replayed.