using a common method for flushes from exe or wb stage
In a single cycle only one of these can have effect on the next pc value. Also previously it was possible that the flush from exe and flush from wb can simultaneously change their epochs. However, now that is not possible. Since the wb flush trumps the exe flush only the wEpoch will toggle if both fire simultaneously. The parent module is respondisble for sending the correct muxed pc from either flushes to the fetch stage.